+++ Packages/com.jonagill.asyncroutines/Runtime/PlayModeTimeProvider.cs
@@ -5,7 +5,7 @@
         public float Time => UnityEngine.Time.time;
         public float DeltaTime => UnityEngine.Time.deltaTime;
         public float FixedDeltaTime => UnityEngine.Time.fixedDeltaTime;
-        public float FrameCount => UnityEngine.Time.frameCount;
+        public int FrameCount => UnityEngine.Time.frameCount;
         public float RealTimeSinceStartup => UnityEngine.Time.realtimeSinceStartup;
     }
 }
