+++ Packages/com.jonagill.asyncroutines/Unit Tests/MockTimeProvider.cs
@@ -5,7 +5,7 @@
         public float Time { get; set; }
         public float DeltaTime { get; set; }
         public float FixedDeltaTime { get; set; }
-        public float FrameCount { get; set; }
+        public int FrameCount { get; set; }
         public float RealTimeSinceStartup { get; set; }
     }    
 }
